src/requirements_server.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
docutils==0.15.2
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
humanfriendly==10.0
ibm-cos-sdk==2.7.0
ibm-cos-sdk-core==2.7.0
ibm-cos-sdk-s3transfer==2.7.0
ibm-watson-machine-learning==1.0.297
idna==3.4
importlib-metadata==6.6.0
jax==0.4.10
jmespath==0.10.0
keras==2.11.0
libclang==16.0.0
lomond==0.3.3
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
numpy==1.23.5
oauthlib==3.2.2
onnx==1.14.0
onnxruntime-gpu==1.14.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
pandas==1.5.3
protobuf==4.23.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tflite-runtime==2.11.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.14.1
zipp==3.15.0
